refactor: 관리자 검증 로직 추가

+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeController.java
@@ -1,7 +1,9 @@
 package roomescape.domain.reservationtime;
 
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.admin.AdminRequestValidator;
 import roomescape.domain.reservationtime.dto.CreateTimeRequest;
 import roomescape.domain.reservationtime.dto.CreateTimeResponse;
 import roomescape.domain.reservationtime.dto.ReservationTimeAvailabilityResponse;
@@ -20,22 +23,35 @@ import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
 public class ReservationTimeController {
 
     private final ReservationTimeService reservationTimeService;
+    private final AdminRequestValidator validator;
 
     @GetMapping("/admin/times")
-    public ResponseEntity<List<ReservationTimeResponse>> getAllReservationTime() {
+    public ResponseEntity<List<ReservationTimeResponse>> getAllReservationTime(HttpServletRequest request) {
+        if (validator.isUnauthorized(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         List<ReservationTimeResponse> response = reservationTimeService.getAllReservationTime();
         return ResponseEntity.ok(response);
     }
 
     @PostMapping("/admin/times")
-    public ResponseEntity<CreateTimeResponse> createReservationTime(@RequestBody CreateTimeRequest request) {
-        request.validate();
-        CreateTimeResponse response = reservationTimeService.createReservationTime(request);
+    public ResponseEntity<CreateTimeResponse> createReservationTime(
+        HttpServletRequest httpServletRequest,
+        @RequestBody CreateTimeRequest createTimeRequest
+    ) {
+        createTimeRequest.validate();
+        if (validator.isUnauthorized(httpServletRequest)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        CreateTimeResponse response = reservationTimeService.createReservationTime(createTimeRequest);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/admin/times/{id}")
-    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteReservationTime(HttpServletRequest request, @PathVariable Long id) {
+        if (validator.isUnauthorized(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         reservationTimeService.deleteReservationTime(id);
         return ResponseEntity.ok().build();
     }
